+++ src/multi-partition.d.ts
@@ -1,0 +1,2 @@
+import { multiPartition } from './index'
+export default multiPartition