+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - pipe
  - splitAt/asyncSplitAt
+ - multiPartition/asyncMultiPartition
 
 ### Changes
  - takeSorted/asyncTakeSorted optional comparator is now the first argument